requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
scrapling==0.3.2
httpx==0.27.2
cachetools==5.5.0
python-dotenv==1.0.1
openai>=1.0.0
